UPDATE: 피드 컴포넌트 분리
