+++ src/Router.js
@@ -17,7 +17,7 @@ const Router = () => {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<Team />} />
+        <Route path="/team" element={<Team />} />
         <Route path="/jisu-Login" element={<JisuLogin />} />
         <Route path="/jisu-Main" element={<JisuMain />} />
 
@@ -25,7 +25,7 @@ const Router = () => {
         <Route path="/sumi-Main" element={<SumiMain />} />
 
         <Route path="/jeoungho-Login" element={<JeounghoLogin />} />
-        <Route path="/jeoungho-Main" element={<JeounghoMain />} />
+        <Route path="/" element={<JeounghoMain />} />
 
         <Route path="/jeongin-Login" element={<JeonginLogin />} />
         <Route path="/jeongin-Main" element={<JeonginMain />} />
