+++ src/pages/shin/Main/JeounghoMain.js
@@ -1,21 +1,15 @@
 import React, { useEffect, useState } from 'react';
-import Comment from '../../../components/Comment/Comment';
-import profile from '../../../assets/shin/KakaoTalk_20230213_204309040.jpg';
-import posted from '../../../assets/shin/KakaoTalk_20230214_161243133.jpg';
+import Feed from '../../../components/Comment/Feed';
 import building from '../../../assets/shin/KakaoTalk_20230216_100337086.jpg';
 import './JeounghoMain.scss';
 
 const JeounghoMain = () => {
   const [isminiMemu, setIsMiniMemu] = useState(false);
-  const [isheart, setIsHeart] = useState(false);
-  const [comment, setComment] = useState('');
-  const [saveComment, setSaveComment] = useState([]);
-  const [numberComments, setNumberComments] = useState(0);
-  const [numberlike, setNumberLike] = useState(0);
   const [idValueFind, setIdValueFind] = useState('');
-  const [foundId, setFoundId] = useState([]);
-  const [isIdArea, setIsIdArea] = useState('hidden');
+  // const [foundId, setFoundId] = useState([]);
+  // const [isIdArea, setIsIdArea] = useState('hidden');
   const [mainData, setMainData] = useState([]);
+  const [mainData2, setMainData2] = useState([]);
 
   useEffect(() => {
     fetch('./data/mainData.json', {
@@ -25,76 +19,32 @@ const JeounghoMain = () => {
       .then(data => [setMainData(data)]);
   }, []);
 
+  useEffect(() => {
+    fetch('./data/feedData.json', {
+      method: 'GET',
+    })
+      .then(res => res.json())
+      .then(data => [setMainData2(data)]);
+  }, []);
+
   const miniMenuToggle = () => {
     setIsMiniMemu(!isminiMemu);
-  };
-
-  const heartToggle = () => {
-    setIsHeart(!isheart);
-  };
-
-  const changeComment = e => {
-    setComment(e.target.value);
   };
 
   const changeIdValueFind = e => {
     setIdValueFind(e.target.value);
   };
 
-  const idReach = idValueFind => {
-    return idValueFind
-      ? (setFoundId(
-          saveComment
-            .filter(item => item.id.includes(idValueFind))
-            .map(item => item.id)
-        ),
-        setIsIdArea(''))
-      : (setFoundId([]), setIsIdArea('hidden'));
-  };
-
-  const commitCreate = e => {
-    return e.key === 'Enter' && comment.length > 0
-      ? (commentInformation(), setComment(''))
-      : '';
-  };
-
-  const postButton = () => {
-    return comment.length > 0 ? (commentInformation(), setComment('')) : '';
-  };
-
-  const commentInformation = () => {
-    const input = {
-      id: random(),
-      value: comment,
-      up: false,
-    };
-    setSaveComment([...saveComment, input]);
-    setNumberComments(numberComments + 1);
-  };
-
-  const likeToggle = item => {
-    saveComment.filter(items => {
-      return items.id === item.id && !item.up
-        ? ((item.up = !item.up), setNumberLike(numberlike + 1))
-        : items.id === item.id && item.up
-        ? ((item.up = !item.up), setNumberLike(numberlike - 1))
-        : item;
-    });
-  };
-
-  const deleteComment = item => {
-    setSaveComment(
-      saveComment.filter(items => {
-        return items.id !== item.id;
-      })
-    );
-    setNumberComments(numberComments - 1);
-    return item.up === true ? setNumberLike(numberlike - 1) : '';
-  };
-
-  const random = () => {
-    return Math.random().toString(36).substr(2, 16);
-  };
+  // const idReach = idValueFind => {
+  //   return idValueFind
+  //     ? (setFoundId(
+  //         saveComment
+  //           .filter(item => item.id.includes(idValueFind))
+  //           .map(item => item.id)
+  //       ),
+  //       setIsIdArea(''))
+  //     : (setFoundId([]), setIsIdArea('hidden'));
+  // };
 
   return (
     <div className="main">
@@ -112,16 +62,16 @@ const JeounghoMain = () => {
               placeholder="검색"
               value={idValueFind}
               onChange={changeIdValueFind}
-              onKeyUp={() => idReach(idValueFind)}
+              // onKeyUp={() => idReach(idValueFind)}
             />
-            <ul className={`searchIdArea${isIdArea}`}>
+            {/* <ul className={`searchIdArea${isIdArea}`}>
               {foundId.map(item => (
                 <li key={item}>
                   <img src="./images/shin/default profile.png" alt="default" />
                   {item}
                 </li>
               ))}
-            </ul>
+            </ul> */}
             <i className="bi bi-search" />
           </div>
           <div className="imgs">
@@ -147,68 +97,11 @@ const JeounghoMain = () => {
       </nav>
 
       <main className="mainContainer">
-        <div className="feeds">
-          <div className="article">
-            <div className="information">
-              <div className="profile">
-                <img src={profile} alt="profile" />
-                <p>hole546</p>
-              </div>
-              <i className="bi bi-three-dots" />
-            </div>
-            <img src={posted} alt="posted" />
-
-            <div className="content">
-              <div className="emotion">
-                <i
-                  onClick={heartToggle}
-                  className={
-                    isheart
-                      ? 'bi bi-suit-heart-fill clicked'
-                      : 'bi bi-suit-heart-fill'
-                  }
-                />
-                <i className="bi bi-chat" />
-                <i className="bi bi-upload" />
-              </div>
-              <i className="bi bi-bookmark" />
-            </div>
-
-            <span className="hearts">좋아요 {numberlike}개</span>
-            <div className="commentSection">
-              <img src={profile} alt="profile" />
-              <p>
-                <strong>hole546님 외 </strong>
-                <strong>{numberComments}명</strong>이 댓글을 다셨습니다.
-              </p>
-            </div>
-            <div className="commentContent">
-              {saveComment.map(item => (
-                <Comment
-                  item={item}
-                  key={item.id}
-                  saveComment={saveComment}
-                  likeToggle={likeToggle}
-                  deleteComment={deleteComment}
-                />
-              ))}
-            </div>
-            <div className="wrap">
-              <input
-                onKeyPress={commitCreate}
-                className="comment"
-                type="text"
-                placeholder="댓글 달기"
-                value={comment}
-                onChange={changeComment}
-              />
-              <button className="postingBtn" onClick={postButton}>
-                게시
-              </button>
-            </div>
-          </div>
+        <div>
+          {mainData2.map(list => (
+            <Feed key={list.id} list={list} />
+          ))}
         </div>
-
         <div className="subMain">
           <div className="article">
             <div className="information">
@@ -223,12 +116,12 @@ const JeounghoMain = () => {
                 <p>스토리</p>
                 <strong>모두 보기</strong>
               </div>
-              {saveComment.map(item => (
+              {/* {saveComment.map(item => (
                 <div className="suggestion" key={item.id}>
                   <img src="./images/shin/default profile.png" alt="default" />
                   <p>{item.id}</p>
                 </div>
-              ))}
+              ))} */}
             </div>
 
             <div className="recommendation">
@@ -236,12 +129,12 @@ const JeounghoMain = () => {
                 <p>회원님을 위한 추천</p>
                 <strong>모두 보기</strong>
               </div>
-              {saveComment.map(item => (
+              {/* {saveComment.map(item => (
                 <div className="suggestion" key={item.id}>
                   <img src="./images/shin/default profile.png" alt="default" />
                   <p>{item.id}</p>
                 </div>
-              ))}
+              ))} */}
             </div>
             <div className="footer">
               {FOOTER.map(item => (
