+++ src/pages/shin/Main/JeounghoMain.scss
@@ -149,7 +149,7 @@
       display: flex;
       justify-content: center;
 
-      width: 40%;
+      width: 100%;
       margin: 10px 10px;
 
       .article {
@@ -304,7 +304,7 @@
       margin: 10px;
 
       .article {
-        width: 90%;
+        width: 60%;
         margin: 10px auto;
 
         .information {
